Changed modID and name to TestMod
